refactor(tuirealm): rename module "value" to "prop_value"
